agent/pyproject.toml
[tool.poetry]
name = "sourcerer"
version = "0.1.0"
description = "Sourcerer - LangGraph Example"
authors = ["Hassan Ijaz <rai.hasan@gmail.com>"]
license = "MIT"

[project]
name = "sourcerer"
version = "0.1.0"
dependencies = [
    "copilotkit[crewai]>=0.1.41",
    "langchain-openai>=0.2.3",
    "langchain-community>=0.3.1",
    "langchain-anthropic>=0.3.1",
    "langchain-google-genai>=2.0.5",
    "langchain>=0.3.4",
    "openai>=1.52.1",
    "motor>=3.4.0",
    "pymongo>=4.7.0",
    "pandas>=2.2.2",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.31.0",
    "requests>=2.32.3",
    "langchain-core>=0.3.25",
    "langgraph-cli[inmem]>=0.1.64",
    "langgraph-checkpoint-sqlite>=2.0.1",
    "aiosqlite>=0.20.0",
    "aiohttp>=3.9.3"
]

[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
copilotkit = {extras = ["crewai"], version = "^0.1.41"}
langchain-openai = "0.2.3"
langchain-community = "^0.3.1"
langchain-anthropic = "0.3.1"
langchain-google-genai = "2.0.5"
langchain = "0.3.4"
openai = "^1.52.1"
motor = "^3.4.0"
pymongo = "^4.7.0"
pandas = "^2.2.2"
pypdf2 = "^3.0.1"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.9"
uvicorn = "^0.31.0"
requests = "^2.32.3"
langchain-core = "^0.3.25"
langgraph-cli = {extras = ["inmem"], version = "^0.1.64"}
langgraph-checkpoint-sqlite = "^2.0.1"
aiosqlite = "^0.20.0"
aiohttp = "^3.9.3"

[tool.poetry.scripts]
demo = "sourcerer.demo:main"
